+++ domain/cart.go
@@ -7,10 +7,10 @@ import (
 )
 
 type Cart struct {
-	ID       int
-	Quantity int
-	// ProductID int
-	// UserID    int
+	ID        int
+	Quantity  int
+	ProductID int
+	UserID    int
 	CreatedAt time.Time
 	UpdatedAt time.Time
 }
@@ -19,23 +19,23 @@ type CartUseCase interface {
 	// AddCart(IDUser int, useCart Cart) (Cart, error)
 	// AddCart(newCart Cart) (Cart, error)
 	GetAllC() ([]Cart, error)
-	// UpCart(IDCart int, updateData Cart) (Cart, error)
-	// DelCart(IDCart int) (bool, error)
-	// GetSpecificCart(cartID int) ([]Cart, error)
+	UpCart(IDCart int, updateData Cart) (Cart, error)
+	DelCart(IDCart int) (bool, error)
+	GetSpecificCart(cartID int) ([]Cart, error)
 }
 
 type CartHandler interface {
 	// InsertCart() echo.HandlerFunc
 	GetAllCart() echo.HandlerFunc
-	// UpdateCart() echo.HandlerFunc
-	// DeleteCart() echo.HandlerFunc
-	// GetCartID() echo.HandlerFunc
+	UpdateCart() echo.HandlerFunc
+	DeleteCart() echo.HandlerFunc
+	GetCartID() echo.HandlerFunc
 }
 
 type CartData interface {
 	// Insert(insertCart Cart) Cart
 	GetAll() []Cart
-	// Update(IDCart int, updatedCart Cart) Cart
-	// Delete(IDCart int) bool
-	// GetCartID(cartID int) []Cart
+	Update(IDCart int, updatedCart Cart) Cart
+	Delete(IDCart int) bool
+	GetCartID(cartID int) []Cart
 }
